MonoDevelop.DesignerSupport.csproj: cleanup code
