+++ main/src/addins/MonoDevelop.DesignerSupport/MonoDevelop.DesignerSupport.csproj
@@ -60,7 +60,6 @@
     <Compile Include="MonoDevelop.DesignerSupport.Toolbox\ItemToolboxNode.cs" />
     <Compile Include="MonoDevelop.DesignerSupport.Toolbox\TextToolboxNode.cs" />
     <Compile Include="MonoDevelop.DesignerSupport.Toolbox\ToolboxItemToolboxNode.cs" />
-    <Compile Include="MonoDevelop.DesignerSupport\IPropertyGrid.cs" />
     <Compile Include="MonoDevelop.DesignerSupport\PropertyPad.cs" />
     <Compile Include="MonoDevelop.DesignerSupport\DesignerSupportService.cs" />
     <Compile Include="MonoDevelop.DesignerSupport\PropertyXwtHostWidget.cs" />
